+++ lib/hangman.ex
