+++ lib/hangman/computer_player.ex
