+++ lib/hangman/game.ex
@@ -1,235 +1,121 @@
-defmodule Hangman.Game do
+ defmodule Hangman.Game do
 
-  @moduledoc """
+  defmodule GameState do
+   #   defstruct word: Hangman.Dictionary.random_word, turnsLeft: 10, correctGuess: [], incorrectGuess: [] ,tempWord: Hangman.Dictionary.random_word
+    defstruct word: "ashwini", turnsLeft: 10, correctGuess: [], incorrectGuess: [], tempWord: "ashwini"
+  end
+ 
 
-  This is the backend for a Hangman game. It manages the game state.
-  Clients make moves, and this code validates them and reports back
-  the updated state.
+   def new_game do
+        r= random_word()
+        IO.puts("")
+        %GameState{ word: r, tempWord: r }
+   end
 
-  We have a well-defined API that is used by client code
-
-  * `game = Hangman.Game.new_game`
-
-    Set up the state for a new game, and return that state. The client
-    applications will pass this state back to your code in all the
-    subsequent API calls.
-
-    The state that's returned will at a minimum contain the word to be
-    guessed.
-
-    As an aid to testing, there's a second form of this function:
-
-    `game = Hangman.Game.new_game(word)`
-
-    This forces `word` to be this game's hidden word.
+   def new_game(word) do
+        game_state = %GameState{ word: String.downcase(word) , tempWord:  String.downcase(word) }
+        game_state
+   end
 
 
-  * `len = Hangman.Game.word_length(game)`
+   def word_length(%{ word: word }) do
+        String.length(word)
+   end
 
-    Return the length of the current word.
+   def letters_used_so_far(state) do
+        state.correctGuess ++ state.incorrectGuess
+   end
+ 
+ 
+    def turns_left(state) do
+        state.turnsLeft
+   end
 
-  * `list = letters_used_so_far(game)`
-
-    The letters that have been guessed so far, returned as a list of
-    single character strings. (This includes both correct and
-    incorrect guessed.
-
-  * `count = turns_left(game)`
-
-    Returns the number of turns remaining before the game is over.
-    For our purposes, a game starts with a generous 10 turns. Each
-    _incorrect_ guess decrements this.
-
-  * `word = word_as_string(game, reveal \\ false)`
-
-     Returns the word to be guessed. If the optional second argument is
-     false, then any unguessed letters will be returned as underscores.
-     If it is true, then the word will be returned complete, showing
-     all letters. Letters and underscores are separated by spaces.
-
-  *  `{game, status, guess} = make_move(game, guess)`
-
-     Accept a guess. Return a three element tuple containing the updated
-     game state, an atom giving the status at the end of the move, and
-     the letter that was guessed.
-
-     The status can be:
-
-     * `:won` — the guess correct and completed the game. The client won
-
-     * `:lost` — the guess was incorrect and the client has run out of
-        turns. The game has been lost.
-
-     * `:good_guess` — the guess occurs one or more times in the word
-
-     * `:bad_guess` — the word does not contain the guess. The number
-       of turns left has been reduced by 1
-
-## Example of use
-
-Here's this module being exercised from an iex session:
-
-    iex(1)> alias Hangman.Game, as: G
-    Hangman.Game
-
-    iex(2)> game = G.new_game
-    . . .
-
-    iex(3)> G.word_length(game)
-    6
-
-    iex(4)> G.word_as_string_string(game)
-    "_ _ _ _ _ _"
-
-    iex(5)> { game, state, guess } = G.make_move(game, "e")
-    . . .
-
-    iex(6)> state
-    :good_guess
-
-    iex(7)> G.word_as_string(game)
-    "_ _ e e _ e"
-
-    iex(8)> { game, state, guess } = G.make_move(game, "q")
-    . . .
-
-    iex(9)> state
-    :bad_guess
-
-    iex(10)> { game, state, guess } = G.make_move(game, "r")
-    . . .
-
-    iex(11)> state
-    :good_guess
-
-    iex(12)> G.word_as_string(game)
-    "_ r e e _ e"
-
-    iex(13)> { game, state, guess } = G.make_move(game, "b")
-    . . .
-    iex(14)> state                                          
-    :bad_guess
-
-    iex(15)> { game, state, guess } = G.make_move(game, "f")
-    . . .
-
-    iex(16)> state
-    :good_guess
-
-    iex(17)> G.word_as_string(game)
-    "f r e e _ e"
-
-    iex(18)> { game, state, guess } = G.make_move(game, "z")
-    . . .
-
-    iex(19)> state
-    :won
-
-    iex(20)> G.word_as_string(game)
-    "f r e e z e"
+   def letters_used_so_far(state) do
+        state.correctGuess ++ state.incorrectGuess
+   end
 
 
-  """
+  def word_as_string_string(state, reveal \\ false) do
+      if(reveal) do
+         state
+      else
+          lst = String.split(state.word, "", trim: true)
+         _str=  Enum.map( lst, fn(x) ->
+            #    IO.puts(x)
+            if Enum.member?(state.correctGuess,to_string(x)) , do: x , else:  "-"   end 
+            )
+          Enum.join(_str,"")
+      end
+    end
 
-  @type state :: map
-  @type ch    :: binary
-  @type optional_ch :: ch | nil
+def make_move(state, guess) do
+ 
+        if(String.contains?(state.tempWord, guess)) do
+            new_state =  %GameState{state | correctGuess: state.correctGuess  ++[guess], turnsLeft: state.turnsLeft - 1 , tempWord: String.replace(state.tempWord, guess, "", global: false)}
+              tuple=  {new_state,  :good_guess, guess }
+            if( String.strip(new_state.tempWord) == ""    ) do 
+                tuple=  {new_state, :won, guess }  
+            end
+ 
+        else
+          new_state=  %GameState{state | incorrectGuess: state.incorrectGuess ++ [guess], turnsLeft: state.turnsLeft - 1}
+           tuple=  {new_state,:bad_guess, guess }
+        
+            if ( new_state.turnsLeft == 0) do 
+            tuple=  {new_state,:lost, guess }
+            end 
+        end
 
-  @doc """
-  Run a game of Hangman with our user. Use the dictionary to
-  find a random word, and then let the user make guesses.
-  """
-
-  @spec new_game :: state
-  def new_game do
+        tuple
   end
 
 
-  @doc """
-  This version of `new_game` doesn't look the word up in the
-  dictionary. Instead, it is passed as a parameter. This is
-  used for testing
-  """
-  @spec new_game(binary) :: state
-  def new_game(word) do
-  end
 
 
-  @doc """
-  `{game, status, guess} = make_move(game, guess)`
-
-  Accept a guess. Return a three element tuple containing the updated
-  game state, an atom giving the status at the end of the move, and
-  the letter that was guessed.
-
-  The status can be:
-
-  * `:won` — the guess correct and completed the game. The client won
-
-  * `:lost` — the guess was incorrect and the client has run out of
-     turns. The game has been lost.
-
-  * `:good_guess` — the guess occurs one or more times in the word
-
-  * `:bad_guess` — the word does not contain the guess. The number
-     of turns left has been reduced by 1
+@moduledoc """
+  We act as an interface to a wordlist (whose name is hardwired in the
+  module attribute `@word_list_file_name`). The list is formatted as
+  one word per line.
   """
 
-  @spec make_move(state, ch) :: { state, atom, optional_ch }
-  def make_move(state, guess) do
-  end
-
+  @word_list_file_name "../../assets/words.8800"
 
   @doc """
-  `len = Hangman.Game.word_length(game)`
-
-  Return the length of the current word.
+  Return a random word from our word list. Whitespace and newlines
+  will have been removed.
   """
-  @spec word_length(state) :: integer
-  def word_length(%{ word: word }) do
+
+  @spec random_word() :: binary
+  def random_word do
+    word_list
+    |> Enum.random
+    |> String.trim
   end
 
   @doc """
-  `list = letters_used_so_far(game)`
-
-  The letters that have been guessed so far, returned as a list of
-  single character strings. (This includes both correct and
-  incorrect guessed.
+  Return a list of all the words in our word list of a given length.
+  Whitespace and newlines will have been removed.
   """
 
-  @spec letters_used_so_far(state) :: [ binary ]
-  def letters_used_so_far(state) do
+  @spec words_of_length(integer)  :: [ binary ]
+  def words_of_length(len) do
+    word_list
+    |> Stream.map(&String.trim/1)
+    |> Enum.filter(&(String.length(&1) == len))
   end
 
-  @doc """
-  `count = turns_left(game)`
-
-  Returns the number of turns remaining before the game is over.
-  For our purposes, a game starts with a generous 10 turns. Each
-  _incorrect_ guess decrements this.
-  """
-
-  @spec turns_left(state) :: integer
-  def turns_left(state) do
-  end
-
-  @doc """
-  `word = word_as_string(game, reveal \\ false)`
-
-  Returns the word to be guessed. If the optional second argument is
-  false, then any unguessed letters will be returned as underscores.
-  If it is true, then the word will be returned complete, showing
-  all letters. Letters and underscores are separated by spaces.
-  """
-
-  @spec word_as_string(state, boolean) :: binary
-  def word_as_string(state, reveal \\ false) do
-  end
 
   ###########################
-  # end of public interface #
+  # End of public interface #
   ###########################
 
-  # Your private functions go here
+  defp word_list do
+    @word_list_file_name
+    |> File.open!
+    |> IO.stream(:line)
+  end
 
- end
+
+
+end
