+++ lib/hangman/human_player.ex
